coś już tam działa

+++ Core/Inc/motors.h
@@ -7,7 +7,7 @@ enum motors_t {LEFT_MOTOR = 0, RIGHT_MOTOR = 1, GO_FORWARD = 2, GO_BACKWARDS = 3
 
 void motors_init(TIM_HandleTypeDef *timer_1, TIM_HandleTypeDef *timer_2);
 
-void motors_set_speed(motors_t motor, uint8_t speed);
+void motors_set_speed(motors_t motor, int16_t speed);
 
 void motors_set_direction(motors_t motor, motors_t direction);
